% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft_module.R
\name{craft_module}
\alias{craft_module}
\title{Craft a C module}
\usage{
craft_module(name = NULL)
}
\arguments{
\item{name}{Creates and opens \verb{src/<name>.h} and \verb{src/<name>.c}.}

\item{header_guards}{Whether to include header guards in \verb{src/<name>.h}.}
}
\value{
Run for its side effect of writing \verb{src/<name>.h} and \verb{src/<name>.c}.
}
\description{
This function creates the initial placeholder for a header file (\code{.h}) and a
C source file (\code{.c}) that together form a module.
}
